Properly abstract away the Chrome-only implementation
